Struct coresimd::arch::x86::__m256i

#[repr(simd)]

pub struct __m256i(_, _, _, _);

This is supported on x86 only.

256-bit wide integer vector type, x86-specific

This type is the same as the __m256i type defined by Intel, representing a 256-bit SIMD register. Usage of this type typically corresponds to the avx and up target features for x86/x86_64.

Internally this type may be viewed as:

i8x32 - thirty two i8 variables packed together

i16x16 - sixteen i16 variables packed together

i32x8 - eight i32 variables packed together

i64x4 - four i64 variables packed together

(as well as unsigned versions). Each intrinsic may interpret the internal bits differently, check the documentation of the intrinsic to see how it's being used.

Note that this means that an instance of __m256i typically just means a "bag of bits" which is left up to interpretation at the point of use.

Examples

#[cfg(target_arch = "x86")]
use std::arch::x86::*;
#[cfg(target_arch = "x86_64")]
use std::arch::x86_64::*;

let all_bytes_zero = _mm256_setzero_si256();
let all_bytes_one = _mm256_set1_epi8(1);
let eight_i32 = _mm256_set_epi32(1, 2, 3, 4, 5, 6, 7, 8);