Function coresimd::vendor::_xsavec64

pub unsafe fn _xsavec64(mem_addr: *mut u8, save_mask: u64)

Perform a full or partial save of the enabled processor states to memory at mem_addr.

xsavec differs from xsave in that it uses compaction and that it may use init optimization. State is saved based on bits [62:0] in save_mask and XCR0. mem_addr must be aligned on a 64-byte boundary.