Struct coresimd::vendor::__m256

#[repr(simd)]

pub struct __m256(_, _, _, _, _, _, _, _);

256-bit wide set of eight f32 types, x86-specific

This type is the same as the __m256 type defined by Intel, representing a 256-bit SIMD register which internally is consisted of eight packed f32 instances. Usage of this type typically corresponds to the avx and up target features for x86/x86_64.

Note that unlike __m256i, the integer version of the 256-bit registers, this __m256 type has one interpretation. Each instance of __m256 always corresponds to f32x8, or eight f32 types packed together.

Most intrinsics using __m256 are prefixed with _mm256_ and are suffixed with "ps" (or otherwise contain "ps"). Not to be confused with "pd" which is used for __m256d.

Examples

use stdsimd::vendor::*;

let eight_zeros = _mm256_setzero_ps();
let eight_ones = _mm256_set1_ps(1.0);
let eight_floats = _mm256_set_ps(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0);